added advisors review to arbitrary generator
