+++ vit-servicing-station-tests/src/common/data/generator/arbitrary/mod.rs
@@ -1,0 +1,5 @@
+mod generator;
+mod snapshot_generator;
+
+pub use generator::ArbitraryGenerator;
+pub use snapshot_generator::ArbitrarySnapshotGenerator;
