+++ vit-servicing-station-tests/src/common/data/generator/mod.rs
@@ -2,6 +2,6 @@ mod arbitrary;
 mod snapshot;
 mod voting;
 
-pub use arbitrary::ArbitraryGenerator;
+pub use arbitrary::{ArbitraryGenerator, ArbitrarySnapshotGenerator};
 pub use snapshot::Snapshot;
 pub use voting::*;
