+++ vit-servicing-station-tests/src/common/data/generator/voting/mod.rs
@@ -5,5 +5,5 @@ pub use plan::{ValidVotePlanGenerator, ValidVotePlanParameters};
 pub use template::{
     parse_challenges, parse_funds, parse_proposals, ArbitraryValidVotingTemplateGenerator,
     ChallengeTemplate, ExternalValidVotingTemplateGenerator, FundTemplate, ProposalTemplate,
-    TemplateLoadError, ValidVotingTemplateGenerator,
+    TemplateLoad, ValidVotingTemplateGenerator,
 };
